Add syncing of data directly from eActivities

+++ collection/lib/crud/products.ts
@@ -1,6 +1,7 @@
 "use server";
 
-import { isValidAcademicYear } from "@docsoc/eactivities";
+import { auth } from "@/auth";
+import { isValidAcademicYear, Product } from "@docsoc/eactivities";
 import { RootItem } from "@prisma/client";
 import { revalidatePath } from "next/cache";
 
@@ -14,6 +15,9 @@ export const getProductsByAcademicYear = async (): Promise<Record<string, RootIt
             academicYear: true,
             id: true,
             name: true,
+            eActivitiesId: true,
+            eActivitiesName: true,
+            eActivitiesURL: true,
         },
     });
 
@@ -38,6 +42,9 @@ export async function getProductsAndVariantByAcademicYearWithCounts(): Promise<
                 select: {
                     id: true,
                     name: true,
+                    eActivitiesId: true,
+                    eActivitiesName: true,
+                    eActivitiesURL: true,
                     Variant: {
                         select: {
                             id: true,
@@ -50,10 +57,10 @@ export async function getProductsAndVariantByAcademicYearWithCounts(): Promise<
                         },
                     },
                 },
+                orderBy: {
+                    id: "asc",
+                },
             },
-        },
-        orderBy: {
-            createdAt: "desc",
         },
     });
 }
@@ -63,6 +70,9 @@ export interface ProductsAndVariantsByAcademicYear {
     RootItem: {
         id: number;
         name: string;
+        eActivitiesId: number | null;
+        eActivitiesName: string | null;
+        eActivitiesURL: string | null;
         Variant: {
             id: number;
             variantName: string;
@@ -74,6 +84,16 @@ export interface ProductsAndVariantsByAcademicYear {
 }
 
 export async function addProducts(academicYear: string, products: string[]): Promise<StatusReturn> {
+    // Action so auth needed
+    const session = await auth();
+
+    if (!session) {
+        return {
+            status: "error",
+            error: "Unauthorized",
+        };
+    }
+
     if (products.length === 0) {
         return {
             status: "error",
@@ -121,6 +141,16 @@ export async function addProducts(academicYear: string, products: string[]): Pro
 }
 
 export async function deleteProduct(productId: number): Promise<StatusReturn> {
+    // Action so auth needed
+    const session = await auth();
+
+    if (!session) {
+        return {
+            status: "error",
+            error: "Unauthorized",
+        };
+    }
+
     try {
         await prisma.rootItem.delete({
             where: {
@@ -140,4 +170,64 @@ export async function deleteProduct(productId: number): Promise<StatusReturn> {
     return {
         status: "success",
     };
+}
+
+export async function updateProductWithEActivitesMetadata({
+    productID,
+    eActivitiesID,
+    eActivitiesData,
+}: {
+    productID: number;
+    eActivitiesID: number;
+    eActivitiesData: Product;
+}): Promise<StatusReturn> {
+    // Action so auth needed
+    const session = await auth();
+
+    if (!session) {
+        return {
+            status: "error",
+            error: "Unauthorized",
+        };
+    }
+
+    if (!eActivitiesData) {
+        return {
+            status: "error",
+            error: "No eActivities data provided",
+        };
+    }
+    try {
+        await prisma.rootItem.update({
+            where: {
+                id: productID,
+            },
+            data: {
+                eActivitiesId: eActivitiesID,
+                eActivitiesName: eActivitiesData.Name,
+                eActivitiesURL: eActivitiesData.URL,
+            },
+        });
+    } catch (e: any) {
+        return {
+            status: "error",
+            error: e.message ?? e.toString(),
+        };
+    }
+
+    revalidatePath("/products");
+
+    return {
+        status: "success",
+    };
+}
+
+export async function getSyncableProducts(): Promise<RootItem[]> {
+    return await prisma.rootItem.findMany({
+        where: {
+            eActivitiesId: {
+                not: null,
+            },
+        },
+    });
 }
